+++ Debug/Editor/GenerateMenuItems.cs
@@ -179,7 +179,7 @@ namespace UnityEditor.ProBuilder
             // *Important* The `instance.enabled` check is redundant for MenuItems, but not for ShortcutManager
             // shortcuts, which atm only have the context of what EditorWindow is active.
             sb.AppendLine( "\t\t\tif(instance != null && instance.enabled)");
-            sb.AppendLine( "\t\t\t\tEditorUtility.ShowNotification(instance.DoAction().notification);");
+            sb.AppendLine( "\t\t\t\tEditorUtility.ShowNotification(instance.PerformAction().notification);");
             sb.AppendLine( "\t\t}");
         }