+++ Editor/EditorCore/PolyShapeTool.cs
@@ -2,8 +2,15 @@
 using UnityEngine;
 using UnityEngine.ProBuilder;
 using UnityEngine.ProBuilder.MeshOperations;
+
 using Math = UnityEngine.ProBuilder.Math;
 using UObject = UnityEngine.Object;
+
+#if !UNITY_2020_2_OR_NEWER
+using ToolManager = UnityEditor.EditorTools.EditorTools;
+#else
+using ToolManager = UnityEditor.EditorTools.ToolManager;
+#endif
 
 namespace UnityEditor.ProBuilder
 {
@@ -22,6 +29,12 @@ namespace UnityEditor.ProBuilder
         const float k_HandleSize = .05f;
 
         GUIContent m_OverlayTitle;
+
+        GUIContent m_IconContent;
+        public override GUIContent toolbarIcon
+        {
+            get { return m_IconContent; }
+        }
 
         Plane m_Plane = new Plane(Vector3.up, Vector3.zero);
 
@@ -84,21 +97,53 @@ namespace UnityEditor.ProBuilder
         {
             m_OverlayTitle = new GUIContent("Poly Shape Tool");
 
+            m_IconContent = new GUIContent()
+            {
+                image = IconUtility.GetIcon("Tools/PolyShape/CreatePolyShape"),
+                text = "Create PolyShape",
+                tooltip = "Create PolyShape"
+            };
+
+#if !UNITY_2020_2_OR_NEWER
+            ToolManager.activeToolChanged += OnToolChanged;
+#endif
+            ProBuilderEditor.selectModeChanged += OnSelectModeChanged;
+            MeshSelection.objectSelectionChanged += OnObjectSelectionChanged;
             Undo.undoRedoPerformed += UndoRedoPerformed;
         }
 
         void OnDisable()
         {
+            SetPolyEditMode(PolyShape.PolyEditMode.None);
+#if !UNITY_2020_2_OR_NEWER
+            ToolManager.activeToolChanged -= OnToolChanged;
+#endif
+            ProBuilderEditor.selectModeChanged -= OnSelectModeChanged;
+            MeshSelection.objectSelectionChanged -= OnObjectSelectionChanged;
             Undo.undoRedoPerformed -= UndoRedoPerformed;
         }
 
-        public void End()
+#if !UNITY_2020_2_OR_NEWER
+        void End()
+#else
+        public override void OnWillBeDeactivated()
+#endif
         {
             if(polygon != null && polygon.polyEditMode != PolyShape.PolyEditMode.None)
                 SetPolyEditMode(PolyShape.PolyEditMode.None);
-            else
-                DestroyImmediate(this);
+
+            DestroyImmediate(this);
         }
+
+#if !UNITY_2020_2_OR_NEWER
+        void OnToolChanged()
+        {
+            if(!ToolManager.IsActiveTool(this))
+                End();
+            else
+                SetPolyEditMode(PolyShape.PolyEditMode.Edit);
+        }
+#endif
 
         /// <summary>
         /// Main GUI update for the tool, calls every secondary methods to place points, update lines and compute the cut
@@ -109,14 +154,14 @@ namespace UnityEditor.ProBuilder
             Event evt = Event.current;
             SceneViewOverlay.Window( m_OverlayTitle, OnOverlayGUI, 0, SceneViewOverlay.WindowDisplayOption.OneWindowPerTitle );
 
-            if (polygon.polyEditMode == PolyShape.PolyEditMode.None)
-                return;
-
             if (polygon == null)
             {
                 SetPolyEditMode(PolyShape.PolyEditMode.None);
                 return;
             }
+
+            if (polygon.polyEditMode == PolyShape.PolyEditMode.None)
+                return;
 
             // used when finishing a loop by clicking the first created point
             if (m_NextMouseUpAdvancesMode && evt.type == EventType.MouseUp)
@@ -202,12 +247,17 @@ namespace UnityEditor.ProBuilder
 
                 case PolyShape.PolyEditMode.Edit:
                 {
-                    if (GUILayout.Button("Quit Editing", UI.EditorGUIUtility.GetActiveStyle("Button")))
+                    if(GUILayout.Button("Quit Editing", UI.EditorGUIUtility.GetActiveStyle("Button")))
+                    {
                         SetPolyEditMode(PolyShape.PolyEditMode.None);
+                        DestroyImmediate(this);
+                    }
                     EditorGUILayout.HelpBox("Move Poly Shape points to update the shape\nPress 'Enter' or 'Space' to Finalize", MessageType.Info);
                     break;
                 }
             }
+
+            //EditorSnapSettings.gridSnapEnabled = EditorGUILayout.Toggle("Use Grid Snapping", EditorSnapSettings.gridSnapEnabled);
 
             EditorGUI.BeginChangeCheck();
 
@@ -274,10 +324,8 @@ namespace UnityEditor.ProBuilder
                 RebuildPolyShapeMesh(polygon);
 
                 //Dirty the polygon for serialization (fix for transition between prefab and scene mode)
-                UnityEditor.EditorUtility.SetDirty(polygon);
-
-                if(mode == PolyShape.PolyEditMode.None)
-                    DestroyImmediate(this);
+                if(polygon != null)
+                    UnityEditor.EditorUtility.SetDirty(polygon);
             }
         }
 
@@ -538,6 +586,14 @@ namespace UnityEditor.ProBuilder
                 m_SelectedIndex = -1;
             }
 
+            if(evt.type == EventType.Repaint && polygon.polyEditMode == PolyShape.PolyEditMode.Path)
+            {
+                Vector3 currentPos = polygon.transform.TransformPoint(m_CurrentPosition);
+                Handles.color = k_HandleColor;
+                Handles.DotHandleCap(-1, currentPos, Quaternion.identity, HandleUtility.GetHandleSize(currentPos) * k_HandleSize, evt.type);
+                Handles.color = Color.white;
+            }
+
             if (polygon.polyEditMode == PolyShape.PolyEditMode.Height)
             {
                 if (!used && evt.type == EventType.MouseUp && evt.button == 0 && !EditorHandleUtility.IsAppendModifier(evt.modifiers))
@@ -796,6 +852,33 @@ namespace UnityEditor.ProBuilder
         {
             if (polygon != null && polygon.polyEditMode != PolyShape.PolyEditMode.None)
                 RebuildPolyShapeMesh(polygon);
+        }
+
+        void OnSelectModeChanged(SelectMode selectMode)
+        {
+            // User changed select mode manually, remove InputTool flag
+            if (polygon != null
+                && polygon.polyEditMode != PolyShape.PolyEditMode.None
+                && !selectMode.ContainsFlag(SelectMode.InputTool))
+            {
+                SetPolyEditMode(PolyShape.PolyEditMode.None);
+            }
+        }
+
+        private void OnObjectSelectionChanged()
+        {
+            if(polygon == null)
+                return;
+
+            if(MeshSelection.activeMesh)
+            {
+                PolyShape shape = MeshSelection.activeMesh.GetComponent<PolyShape>();
+                if(shape != null && shape != polygon)
+                {
+                    //Quit Polygon edit mode and deactivate the tool
+                    SetPolyEditMode(PolyShape.PolyEditMode.None);
+                }
+            }
         }
     }
 }