+++ Editor/MenuActions/Geometry/ExtrudeEdges.cs
@@ -53,10 +53,10 @@ namespace UnityEditor.ProBuilder.Actions
             GUILayout.FlexibleSpace();
 
             if (GUILayout.Button("Extrude Edges"))
-                DoAction();
+                PerformAction();
         }
 
-        public override ActionResult DoAction()
+        protected override ActionResult PerformActionImplementation()
         {
             if (MeshSelection.selectedObjectCount < 1)
                 return ActionResult.NoSelection;