+++ Editor/MenuToggles/Geometry/CutToolToggle.cs
@@ -30,7 +30,7 @@ namespace UnityEditor.ProBuilder.Actions
 
         public override SelectMode validSelectModes
         {
-            get { return SelectMode.Vertex | SelectMode.Edge | SelectMode.Face | SelectMode.Object; }
+            get { return SelectMode.Vertex | SelectMode.Edge | SelectMode.Face; }
         }
 
         protected override bool hasFileMenuEntry
@@ -50,15 +50,14 @@ namespace UnityEditor.ProBuilder.Actions
             get => MeshSelection.selectedObjectCount > 0;
         }
 
-        internal override ActionResult StartActivation()
+        protected override ActionResult PerformActionImplementation()
         {
-            ProBuilderEditor.selectMode = SelectMode.Object;
-
             m_Tool = ScriptableObject.CreateInstance<CutTool>();
             ToolManager.SetActiveTool(m_Tool);
 
             Undo.RegisterCreatedObjectUndo(m_Tool, "Open Cut Tool");
 
+            MenuAction.onPerformAction += ActionPerformed;
             ToolManager.activeToolChanging += LeaveTool;
             ProBuilderEditor.selectModeChanged += OnSelectModeChanged;
 
@@ -70,12 +69,21 @@ namespace UnityEditor.ProBuilder.Actions
 
         internal override ActionResult EndActivation()
         {
+            MenuAction.onPerformAction -= ActionPerformed;
             ToolManager.activeToolChanging -= LeaveTool;
             ProBuilderEditor.selectModeChanged -= OnSelectModeChanged;
 
             Object.DestroyImmediate(m_Tool);
 
+            ProBuilderEditor.instance.Repaint();
+
             return new ActionResult(ActionResult.Status.Success,"Cut Tool Ends");
+        }
+
+        void ActionPerformed(MenuAction newActionPerformed)
+        {
+            if(ToolManager.IsActiveTool(m_Tool) && newActionPerformed.GetType() != this.GetType())
+                LeaveTool();
         }
 
         void OnSelectModeChanged(SelectMode obj)