+++ Tests/Editor/Object/CopyPaste.cs
@@ -5,13 +5,14 @@ using NUnit.Framework;
 using UnityEngine.ProBuilder.Tests.Framework;
 using UnityEditor.ProBuilder;
 using UnityEngine.ProBuilder;
+using UnityEngine.ProBuilder.Shapes;
 
 static class CopyPaste
 {
     [Test]
     public static void CopyWithVerify_CreatesNewMeshAsset()
     {
-        var original = ShapeGenerator.CreateShape(ShapeType.Cube);
+        var original = ShapeFactory.Instantiate<Cube>();
         var copy = UObject.Instantiate(original);
 
         try
@@ -34,7 +35,7 @@ static class CopyPaste
     [Test]
     public static void Copy_ReferencesOriginalMesh()
     {
-        var original = ShapeGenerator.CreateShape(ShapeType.Cube);
+        var original = ShapeFactory.Instantiate<Cube>();
         var copy = UObject.Instantiate(original);
 
         try