+++ Tests/Framework/TestUtility.cs
@@ -15,6 +15,7 @@ using PackageInfo = UnityEditor.PackageManager.PackageInfo;
 #endif
 using UnityEngine.ProBuilder.MeshOperations;
 using UnityEngine.SceneManagement;
+using UnityEngine.ProBuilder.Shapes;
 
 namespace UnityEngine.ProBuilder.Tests.Framework
 {
@@ -96,6 +97,7 @@ namespace UnityEngine.ProBuilder.Tests.Framework
             {
                 var shapes = Enum.GetValues(typeof(ShapeType)) as ShapeType[];
                 ProBuilderMesh[] primitives = new ProBuilderMesh[shapes.Length];
+
                 for (int i = 0, c = shapes.Length; i < c; i++)
                 {
                     primitives[i] = ShapeGenerator.CreateShape(shapes[i]);
@@ -517,7 +519,7 @@ namespace UnityEngine.ProBuilder.Tests.Framework
 
         public static SimpleTuple<ProBuilderMesh, Face> CreateCubeWithNonContiguousMergedFace()
         {
-            var cube = ShapeGenerator.CreateShape(ShapeType.Cube);
+            var cube = ShapeFactory.Instantiate<Cube>();
 
             Assume.That(cube, Is.Not.Null);