+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [5.0.0-pre.8] - 2020-12-23
+
+### Internal
+
+- Remove backwards compatibility breaking API changes.
+- Fix `Material.SetInt` deprecation warnings.
+- Fix failing `Undo` tests on macOS.
 
 ### Features
 
@@ -17,12 +23,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Bug Fixes
 - [case: 1304442] Update package description for SRPs : warning to users to add samples projects.
+- [case: 1300329] Fixing Undo completely reverting all actions.
 - [case: 1299638] Fixed missing dependency on com.unity.modules.physics.
 - [case: 1296104] Fixing freeze transform with negative scales.
 - [case: 1296494] Fixing audio visualizer sample.
 - [case: 1296428] Cleaning Polyshape tool when leaving the EditorTool.
 - [case: 1296427] Removing NullRef Exception on Undo/redo with CutTool.
 - [case: 1296422] Closing MenuToolToggles actions when closing ProBuilder window.
+- [case: 1296520] Fixing `New Poly Shape` menu entry not working
 - [case: 1254339] Correct offset when rendering UVs and correct export when UV Editor is docked.
 - Fixed PolyShape in Prefab mode: PolyShapeMode was not serialized when exiting prefab mode. Update point insertion visualization.
 - [case: 1259845] Fixed dimension overlay being hidden on playmode or reboot of the editor.